@@ -4,13 +4,13 @@ description: Demonstrates how to use the image_crop plugin.
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.0.0-dev.68.0 <3.0.0"
+  sdk: ">=2.17.0 <4.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
 
-  cupertino_icons: ^0.1.3
+  cupertino_icons: ^1.0.2
   image_picker: ^0.7.4
 
 dev_dependencies:
